requirements.txt
allure-python-commons~=2.14.3
selenium~=4.33.0
pytest~=8.4.1
requests~=2.32.4
Faker~=37.4.2